backend/requirements.txt
chunking_evaluation==0.1.0
faiss_cpu==1.10.0
fastapi==0.115.12
ipython==8.12.3
langchain_community==0.3.21
langchain_core==0.3.54
langchain_experimental==0.3.4
langchain_openai==0.3.14
langchain_text_splitters==0.3.8
neo4j==5.28.1
numpy==1.26.4     
pdf2image==1.17.0
Pillow==11.2.1
pydantic==2.11.3
PyPDF2==3.0.1
python-dotenv==1.1.0
requests==2.32.3
scikit_learn==1.4.2   
semantic_chunkers==0.0.10
semantic_router==0.1.7
spacy==3.8.4
supabase==2.15.0
tqdm==4.66.4
unstructured==0.17.2
